requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
pyyaml>=6.0
numpy>=1.26.0
opencv-python>=4.10.0
ultralytics>=8.4.0
pydantic>=2.8.0
